Get rid of model_import dependency from pipeline_types
