+++ crates/model_import/src/lib.rs
@@ -2,7 +2,7 @@ use std::{f32::consts::PI, path::PathBuf, sync::Arc};
 
 use ambient_animation::AnimationOutputs;
 use ambient_core::{bounding::local_bounding_aabb, transform::translation};
-use ambient_editor_derive::ElementEditor;
+use ambient_pipeline_types::models::{MaterialFilter, ModelTextureSize, ModelTransform};
 use ambient_renderer::materials::pbr_material::PbrMaterialDesc;
 use ambient_std::{
     asset_cache::{AssetCache, SyncAssetKeyExt},
@@ -16,7 +16,6 @@ use glam::{Mat4, Vec3, Vec4};
 use image::RgbaImage;
 use model_crate::{ModelCrate, ModelNodeRef};
 use relative_path::RelativePathBuf;
-use serde::{Deserialize, Serialize};
 
 pub(crate) mod animation_bind_id;
 pub mod assimp;
@@ -106,37 +105,6 @@ impl ModelImportPipeline {
     // }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ElementEditor)]
-#[serde(tag = "type")]
-pub enum MaterialFilter {
-    /// Replace all materials.
-    All,
-    /// Replace all materials that match this name exactly.
-    ByName {
-        /// The material name to replace. Must match exactly (i.e. is case-sensitive and does not ignore whitespace).
-        name: String,
-    },
-}
-impl MaterialFilter {
-    pub fn by_name(name: impl Into<String>) -> Self {
-        Self::ByName { name: name.into() }
-    }
-    fn matches(&self, mat: &PbrMaterialDesc) -> bool {
-        match self {
-            MaterialFilter::All => true,
-            MaterialFilter::ByName { name } => mat.name.as_ref() == Some(name),
-        }
-    }
-    fn is_all(&self) -> bool {
-        matches!(self, MaterialFilter::All)
-    }
-}
-impl Default for MaterialFilter {
-    fn default() -> Self {
-        Self::All
-    }
-}
-
 #[derive(Clone, Debug)]
 pub enum ModelImportTransform {
     ImportModelFromUrl {
@@ -220,7 +188,9 @@ impl ModelImportTransform {
                     .model_world_mut()
                     .add_resource(ambient_core::name(), name.clone());
             }
-            ModelImportTransform::Transform(transform) => transform.apply(model_crate),
+            ModelImportTransform::Transform(transform) => {
+                apply_model_transform(transform, model_crate)
+            }
             ModelImportTransform::OverrideMaterial { filter, material } => {
                 model_crate.override_material(filter, (**material).clone());
             }
@@ -251,169 +221,87 @@ impl ModelImportTransform {
     }
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(tag = "type")]
-pub enum ModelTransform {
-    /// Rotate Y up to Z up.
-    RotateYUpToZUp,
-    /// Rotate X by `deg` degrees.
-    RotateX {
-        /// The degrees to rotate this model around the X axis.
-        deg: f32,
-    },
-    /// Rotate Y by `deg` degrees.
-    RotateY {
-        /// The degrees to rotate this model around the Y axis.
-        deg: f32,
-    },
-    /// Rotate Z by `deg` degrees.
-    RotateZ {
-        /// The degrees to rotate this model around the Z axis.
-        deg: f32,
-    },
-    /// Scale this model.
-    Scale {
-        /// The factor to scale this model by.
-        scale: f32,
-    },
-    /// Translate this model.
-    Translate {
-        /// The translation to apply to this model (i.e. this model will be moved by `translation` in the current coordinate space).
-        translation: Vec3,
-    },
-    /// Scale this model's AABB.
-    ScaleAABB {
-        /// The factor to scale this model's AABB by.
-        scale: f32,
-    },
-    /// Scale this model's animations (spatially, not in time).
-    ScaleAnimations {
-        /// The factor to scale this model's animations by.
-        scale: f32,
-    },
-    /// Re-root this mesh.
-    SetRoot {
-        /// The name of the node to set as the new root for this mesh.
-        name: String,
-    },
-    /// Re-center this mesh such that the root is located at the origin.
-    Center,
-}
-impl ModelTransform {
-    pub fn apply(&self, model_crate: &mut ModelCrate) {
-        match self {
-            ModelTransform::RotateYUpToZUp => {
-                let transform = Mat4::from_cols(Vec4::X, Vec4::Z, Vec4::Y, Vec4::W);
-                model_crate.model_mut().transform(transform);
-            }
-            ModelTransform::RotateX { deg } => {
-                model_crate
-                    .model_mut()
-                    .transform(Mat4::from_rotation_x(deg * PI / 180.));
-            }
-            ModelTransform::RotateY { deg } => {
-                model_crate
-                    .model_mut()
-                    .transform(Mat4::from_rotation_y(deg * PI / 180.));
-            }
-            ModelTransform::RotateZ { deg } => {
-                model_crate
-                    .model_mut()
-                    .transform(Mat4::from_rotation_z(deg * PI / 180.));
-            }
-            ModelTransform::Scale { scale } => {
-                model_crate
-                    .model_mut()
-                    .transform(Mat4::from_scale(Vec3::ONE * *scale));
-            }
-            ModelTransform::Translate { translation } => {
-                model_crate
-                    .model_mut()
-                    .transform(Mat4::from_translation(*translation));
-            }
-            ModelTransform::ScaleAABB { scale } => {
-                let world = model_crate.model_world_mut();
-                let aabb = world.resource_mut(local_bounding_aabb());
-                aabb.min *= *scale;
-                aabb.max *= *scale;
-            }
-            ModelTransform::ScaleAnimations { scale: anim_scale } => {
-                for clip in model_crate.animations.content.values_mut() {
-                    *clip = clip.map_outputs(|outputs| {
-                        if outputs.component() == translation() {
-                            match outputs {
-                                AnimationOutputs::Vec3 { component, data } => {
-                                    AnimationOutputs::Vec3 {
-                                        component: *component,
-                                        data: data.iter().map(|x| *x * *anim_scale).collect(),
-                                    }
-                                }
-                                AnimationOutputs::Quat {
-                                    component: _,
-                                    data: _,
-                                } => unreachable!(),
-                                AnimationOutputs::Vec3Field {
-                                    component,
-                                    field,
-                                    data,
-                                } => AnimationOutputs::Vec3Field {
-                                    component: *component,
-                                    field: *field,
-                                    data: data.iter().map(|x| *x * *anim_scale).collect(),
-                                },
-                            }
-                        } else {
-                            outputs.clone()
+pub fn apply_model_transform(model_transform: &ModelTransform, model_crate: &mut ModelCrate) {
+    match model_transform {
+        ModelTransform::RotateYUpToZUp => {
+            let transform = Mat4::from_cols(Vec4::X, Vec4::Z, Vec4::Y, Vec4::W);
+            model_crate.model_mut().transform(transform);
+        }
+        ModelTransform::RotateX { deg } => {
+            model_crate
+                .model_mut()
+                .transform(Mat4::from_rotation_x(deg * PI / 180.));
+        }
+        ModelTransform::RotateY { deg } => {
+            model_crate
+                .model_mut()
+                .transform(Mat4::from_rotation_y(deg * PI / 180.));
+        }
+        ModelTransform::RotateZ { deg } => {
+            model_crate
+                .model_mut()
+                .transform(Mat4::from_rotation_z(deg * PI / 180.));
+        }
+        ModelTransform::Scale { scale } => {
+            model_crate
+                .model_mut()
+                .transform(Mat4::from_scale(Vec3::ONE * *scale));
+        }
+        ModelTransform::Translate { translation } => {
+            model_crate
+                .model_mut()
+                .transform(Mat4::from_translation(*translation));
+        }
+        ModelTransform::ScaleAABB { scale } => {
+            let world = model_crate.model_world_mut();
+            let aabb = world.resource_mut(local_bounding_aabb());
+            aabb.min *= *scale;
+            aabb.max *= *scale;
+        }
+        ModelTransform::ScaleAnimations { scale: anim_scale } => {
+            for clip in model_crate.animations.content.values_mut() {
+                *clip = clip.map_outputs(|outputs| {
+                    if outputs.component() == translation() {
+                        match outputs {
+                            AnimationOutputs::Vec3 { component, data } => AnimationOutputs::Vec3 {
+                                component: *component,
+                                data: data.iter().map(|x| *x * *anim_scale).collect(),
+                            },
+                            AnimationOutputs::Quat {
+                                component: _,
+                                data: _,
+                            } => unreachable!(),
+                            AnimationOutputs::Vec3Field {
+                                component,
+                                field,
+                                data,
+                            } => AnimationOutputs::Vec3Field {
+                                component: *component,
+                                field: *field,
+                                data: data.iter().map(|x| *x * *anim_scale).collect(),
+                            },
                         }
-                    });
-                }
+                    } else {
+                        outputs.clone()
+                    }
+                });
             }
-            ModelTransform::SetRoot { name } => {
-                if let Some(id) = model_crate.model().get_entity_id_by_name(name) {
-                    model_crate.make_new_root(id);
-                }
+        }
+        ModelTransform::SetRoot { name } => {
+            if let Some(id) = model_crate.model().get_entity_id_by_name(name) {
+                model_crate.make_new_root(id);
             }
-            ModelTransform::Center => {
-                model_crate.model_mut().center();
-            }
+        }
+        ModelTransform::Center => {
+            model_crate.model_mut().center();
         }
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, ElementEditor)]
-pub enum ModelTextureSize {
-    /// Cap this model's textures to 128x128.
-    X128,
-    /// Cap this model's textures to 256x256.
-    X256,
-    /// Cap this model's textures to 512x512.
-    X512,
-    /// Cap this model's textures to 1024x1024.
-    X1024,
-    /// Cap this model's textures to 2048x2048.
-    X2048,
-    /// Cap this model's textures to 4096x4096.
-    X4096,
-    /// Cap this model's textures to SIZE x SIZE.
-    /// It is strongly recommended that this is a power of two.
-    Custom(u32),
-}
-impl ModelTextureSize {
-    pub fn size(&self) -> u32 {
-        match self {
-            ModelTextureSize::X128 => 128,
-            ModelTextureSize::X256 => 256,
-            ModelTextureSize::X512 => 512,
-            ModelTextureSize::X1024 => 1024,
-            ModelTextureSize::X2048 => 2048,
-            ModelTextureSize::X4096 => 4096,
-            ModelTextureSize::Custom(size) => *size,
-        }
-    }
-}
-impl Default for ModelTextureSize {
-    fn default() -> Self {
-        Self::X512
+fn material_filter_matches(filter: &MaterialFilter, mat: &PbrMaterialDesc) -> bool {
+    match filter {
+        MaterialFilter::All => true,
+        MaterialFilter::ByName { name } => mat.name.as_ref() == Some(name),
     }
 }
