+++ crates/pipeline_types/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 [dependencies]
 ambient_std = { path = "../std", version = "0.3.0-dev" }
 ambient_ecs = { path = "../ecs", version = "0.3.0-dev" }
-ambient_model_import = { path = "../model_import", version = "0.3.0-dev" }
 ambient_gpu = { path = "../gpu", version = "0.3.0-dev" }
 
 serde.workspace = true
