+++ crates/pipeline_types/src/models/mod.rs
@@ -1,8 +1,8 @@
 use ambient_ecs::Entity;
-use ambient_model_import::{MaterialFilter, ModelTextureSize, ModelTransform};
 use serde::{Deserialize, Serialize};
 
 use crate::{is_default, is_false, is_true, materials::PipelinePbrMaterial, true_value};
+use glam::Vec3;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(deny_unknown_fields)]
@@ -143,4 +143,115 @@ pub enum ColliderType {
     TriggerArea,
     /// This object should only be present in the picking scene.
     Picking,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(tag = "type")]
+pub enum MaterialFilter {
+    /// Replace all materials.
+    All,
+    /// Replace all materials that match this name exactly.
+    ByName {
+        /// The material name to replace. Must match exactly (i.e. is case-sensitive and does not ignore whitespace).
+        name: String,
+    },
+}
+impl MaterialFilter {
+    pub fn by_name(name: impl Into<String>) -> Self {
+        Self::ByName { name: name.into() }
+    }
+    pub fn is_all(&self) -> bool {
+        matches!(self, MaterialFilter::All)
+    }
+}
+impl Default for MaterialFilter {
+    fn default() -> Self {
+        Self::All
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum ModelTextureSize {
+    /// Cap this model's textures to 128x128.
+    X128,
+    /// Cap this model's textures to 256x256.
+    X256,
+    /// Cap this model's textures to 512x512.
+    X512,
+    /// Cap this model's textures to 1024x1024.
+    X1024,
+    /// Cap this model's textures to 2048x2048.
+    X2048,
+    /// Cap this model's textures to 4096x4096.
+    X4096,
+    /// Cap this model's textures to SIZE x SIZE.
+    /// It is strongly recommended that this is a power of two.
+    Custom(u32),
+}
+impl ModelTextureSize {
+    pub fn size(&self) -> u32 {
+        match self {
+            ModelTextureSize::X128 => 128,
+            ModelTextureSize::X256 => 256,
+            ModelTextureSize::X512 => 512,
+            ModelTextureSize::X1024 => 1024,
+            ModelTextureSize::X2048 => 2048,
+            ModelTextureSize::X4096 => 4096,
+            ModelTextureSize::Custom(size) => *size,
+        }
+    }
+}
+impl Default for ModelTextureSize {
+    fn default() -> Self {
+        Self::X512
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(tag = "type")]
+pub enum ModelTransform {
+    /// Rotate Y up to Z up.
+    RotateYUpToZUp,
+    /// Rotate X by `deg` degrees.
+    RotateX {
+        /// The degrees to rotate this model around the X axis.
+        deg: f32,
+    },
+    /// Rotate Y by `deg` degrees.
+    RotateY {
+        /// The degrees to rotate this model around the Y axis.
+        deg: f32,
+    },
+    /// Rotate Z by `deg` degrees.
+    RotateZ {
+        /// The degrees to rotate this model around the Z axis.
+        deg: f32,
+    },
+    /// Scale this model.
+    Scale {
+        /// The factor to scale this model by.
+        scale: f32,
+    },
+    /// Translate this model.
+    Translate {
+        /// The translation to apply to this model (i.e. this model will be moved by `translation` in the current coordinate space).
+        translation: Vec3,
+    },
+    /// Scale this model's AABB.
+    ScaleAABB {
+        /// The factor to scale this model's AABB by.
+        scale: f32,
+    },
+    /// Scale this model's animations (spatially, not in time).
+    ScaleAnimations {
+        /// The factor to scale this model's animations by.
+        scale: f32,
+    },
+    /// Re-root this mesh.
+    SetRoot {
+        /// The name of the node to set as the new root for this mesh.
+        name: String,
+    },
+    /// Re-center this mesh such that the root is located at the origin.
+    Center,
 }
